Production Visualization in Command Panel

+++ Assets/Scriptcs/Unit/UnitData.cs
@@ -11,6 +11,6 @@ public class UnitData : ScriptableObject
     public float speed;
     public Sprite unitSprite;
     public GameObject unitPrefab;
-    public int productionTime;
+    public float productionTime;
     [field: SerializeField] public List<ObjectPrices> objectPrices = new List<ObjectPrices>();
 }
